1.5 : Env var
Try to access the application at the URL indicated by the logs. Go back to your terminal and you'll be presented with a nice error stacktrace. This error occurs because the program expects to find the environment variable CURRENT_ENVIRONMENT present in the container.

Restart the container and add the missing environment variable so that the service responds correctly (you can set the value you prefer).

The application starts up correctly and you can access it by following the URL displayed in the logs.