1.6: Let's Ship it
In order to share your fantastic new creation with the world, we're going to push the image to the official Docker registry called Docker Hub (or Docker Store)

To do this:

Create an account on Docker Hub
Create a new repository attached to your account
Then push your newly created image to your repository.
Your image cannot be pushed as it is (:warning:) Why do you think this is? Change the name of your image and push it back onto your Docker Hub repository.